messenger-main/requirements.txt
Flask==3.1.2
Flask-SocketIO==5.5.1
cryptography==46.0.1
python-socketio==5.15.0
python-engineio==4.12.3
eventlet==0.39.1
